changed cross_dispersion variable type from double to int
git-svn-id: svn://136.177.114.72/svn_GW/phast/trunk@485 1feff8c3-07ed-0310-ac33-dd36852eb9cd

+++ srcinput/hstinpt.h
@@ -631,7 +631,7 @@ EXTERNAL int solver_save_directions;
 EXTERNAL int solver_maximum;
 EXTERNAL double solver_space;
 EXTERNAL double solver_time;
-EXTERNAL double cross_dispersion;
+EXTERNAL int cross_dispersion;
 
 /* time stepping */
 EXTERNAL struct time_series time_step;
